Moved test code to it's own file.
